changes in getAl companies and get particular companies

+++ DPUWebProject/src/main/java/com/dpu/controller/CompanyController.java
@@ -138,7 +138,7 @@ public class CompanyController extends MessageProperties {
 	}
 
 	@RequestMapping(value = "/{id}", produces = MediaType.APPLICATION_JSON_VALUE, method = RequestMethod.GET)
-	public Object get(@PathVariable("id") int id) {
+	public Object get(@PathVariable("id") Long id) {
 		
 		logger.info("[get] : Enter : ID : "+id);
 		String json = new String();
