+++ DPUWebProject/src/main/java/com/dpu/service/CompanyBillingLocationService.java
@@ -12,8 +12,8 @@ public interface CompanyBillingLocationService {
 	
 	boolean delete(CompanyBillingLocation companyBillingLocation);
 	
-	List<CompanyBillingLocation> getAll(int companyId);
+	List<CompanyBillingLocation> getAll(Long companyId);
 	
-	CompanyBillingLocation get(int id);
+	CompanyBillingLocation get(Long id);
 	
 }
